refactor: rename LineIF to Line
