+++ src/main/java/nextstep/ladder/Controller.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Line;
+import nextstep.ladder.domain.SingleLine;
 import nextstep.ladder.domain.Spoke;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
@@ -25,7 +25,7 @@ public class Controller {
         resultView.printResult(players, getLines(ladderHeight, players.size()));
     }
 
-    private List<Line> getLines(int ladderHeight, int playersCount) {
+    private List<SingleLine> getLines(int ladderHeight, int playersCount) {
         return IntStream.range(0, ladderHeight)
                 .mapToObj(__ -> Spoke.fromCount(playersCount - 1, RANDOM::nextBoolean))
                 .map(Spoke::toLine)
