--- src/main/java/nextstep/ladder/domain/LineIF.java
@@ -1,6 +1,0 @@
-package nextstep.ladder.domain;
-
-interface LineIF {
-
-    Position moveOn(Position from);
-}
