+++ src/main/java/nextstep/ladder/domain/Lines.java
@@ -4,16 +4,23 @@ import nextstep.ladder.utils.BinaryOperators;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Stream;
 
-class Lines implements LineIF {
+import static java.util.stream.Collectors.toList;
+
+class Lines implements Line {
     private final List<Line> lines;
 
     public Lines(List<Line> lines) {
         this.lines = lines;
     }
 
-    public static LineIF of(Line... lines) {
+    public static Line of(Line... lines) {
         return new Lines(Arrays.asList(lines));
+    }
+
+    public static Line of(Spoke... spokes) {
+        return new Lines(Stream.of(spokes).map(Spoke::toLine).collect(toList()));
     }
 
     @Override
