+++ src/test/java/nextstep/ladder/domain/LineTest.java
@@ -9,21 +9,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class LineTest {
 
-    @DisplayName("오른쪽에 발판이 존재하는지 알 수 있다")
-    @Test
-    void hasSpokeRightSide() {
-        Line line = Spoke.of(true, false, true).toLine();
-        assertAll(
-                () -> assertThat(line.hasSpokeRightSide(Position.of(0))).isTrue(),
-                () -> assertThat(line.hasSpokeRightSide(Position.of(1))).isFalse(),
-                () -> assertThat(line.hasSpokeRightSide(Position.of(2))).isTrue()
-        );
-    }
-
     @DisplayName("|-----| 의 경우 입력 포지션과 출력 포지션이 다르다")
     @Test
     void moveWithLine() {
-        Line line = Spoke.of(true).toLine();
+        Line line = Lines.of(Spoke.of(true));
         assertAll(
                 () -> assertThat(line.moveOn(Position.of(0))).isEqualTo(Position.of(1)),
                 () -> assertThat(line.moveOn(Position.of(1))).isEqualTo(Position.of(0))
@@ -33,7 +22,7 @@ public class LineTest {
     @DisplayName("라인 바깥쪽 포지션은 입력받을 수 없다")
     @Test
     void outOfLineEndException() {
-        Line line = Spoke.of(true).toLine();
+        Line line = Lines.of(Spoke.of(true));
         assertThatThrownBy(() -> line.moveOn(Position.of(2)))
                 .isInstanceOf(OutOfLineException.class);
     }
@@ -41,7 +30,7 @@ public class LineTest {
     @DisplayName("|-----|     | 의 경우 입력 포지션과 출력 포지션이 다르다")
     @Test
     void moveWithLineWithThreePole() {
-        Line line = Spoke.of(true, false).toLine();
+        Line line = Lines.of(Spoke.of(true, false));
         assertAll(
                 () -> assertThat(line.moveOn(Position.of(0))).isEqualTo(Position.of(1)),
                 () -> assertThat(line.moveOn(Position.of(1))).isEqualTo(Position.of(0)),
@@ -52,7 +41,7 @@ public class LineTest {
     @DisplayName("|     |-----| 의 경우 입력 포지션과 출력 포지션이 다르다")
     @Test
     void moveWithLineWithThreePoleOther() {
-        Line line = Spoke.of(false, true).toLine();
+        Line line = Lines.of(Spoke.of(false, true));
         assertAll(
                 () -> assertThat(line.moveOn(Position.of(0))).isEqualTo(Position.of(0)),
                 () -> assertThat(line.moveOn(Position.of(1))).isEqualTo(Position.of(2)),
@@ -64,7 +53,7 @@ public class LineTest {
                  "|-----| 의 경우 입력 포지션과 출력 포지션이 같다")
     @Test
     void moveWithTwoLineTwoStairs() {
-        LineIF line = Lines.of(Spoke.of(true).toLine(), Spoke.of(true).toLine());
+        Line line = Lines.of(Spoke.of(true), Spoke.of(true));
 
         assertAll(
                 () -> assertThat(line.moveOn(Position.of(0))).isEqualTo(Position.of(0)),
@@ -76,7 +65,7 @@ public class LineTest {
                  "|     |-----| 의 경우 입력 포지션과 출력 포지션이 하나씩 교차된다")
     @Test
     void moveWithThreeLineTwoStairs() {
-        LineIF line = Lines.of(Spoke.of(true, false).toLine(), Spoke.of(false, true).toLine());
+        Line line = Lines.of(Spoke.of(true, false).toLine(), Spoke.of(false, true).toLine());
 
         assertAll(
                 () -> assertThat(line.moveOn(Position.of(0))).isEqualTo(Position.of(2)),
